Refactor: Move classes into sub-packages
