+++ src/main/java/jarvis/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package jarvis;
+package jarvis.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
