+++ src/main/java/jarvis/tasks/Event.java
@@ -1,4 +1,4 @@
-package jarvis;
+package jarvis.tasks;
 
 public class Event extends Task {
     protected String from;
