+++ src/test/java/jarvis/TodoTest.java
@@ -1,5 +1,6 @@
 package jarvis;
 
+import jarvis.tasks.Todo;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
